Refactor: split network.php into network/ namespace files
network.php had grown to 624 lines after JSON Feed support landed. Split
the feed-ingestion concerns into src/network/ (all in Lamb\Network, the
same namespace-split pattern as response/ and theme/):

- network/sources.php   feed config, SimplePie setup, RSS/Atom crawl recording
- network/json_feed.php JSON Feed detection, parsing, crawl, JsonFeedItem adapter
- network/ingest.php    item -> post pipeline (dedup, slug, citation)
- network/status.php    the feedstatus health bean (read by the Logs tab)

network.php is now the thin cron orchestrator (114 lines). Extracted the
inline SimplePie construction into init_simplepie_feed() and the JSON/XML
dispatch into crawl_feed(), which collapses the duplicated OK/FAILED summary
printf into one place. Behaviour is unchanged.

995 unit tests pass; composer lint and analyse clean.
