+++ composer.json
@@ -45,6 +45,10 @@
       "src/lamb.php",
       "src/routes.php",
       "src/network.php",
+      "src/network/status.php",
+      "src/network/sources.php",
+      "src/network/ingest.php",
+      "src/network/json_feed.php",
       "src/post.php",
       "src/response.php",
       "src/response/auth.php",
